+++ internal/configs/parser_config_dir_test.go
@@ -12,6 +12,8 @@ import (
 	"testing"
 
 	"github.com/hashicorp/hcl/v2"
+	"github.com/opentofu/opentofu/internal/addrs"
+	"github.com/zclconf/go-cty/cty"
 )
 
 // TestParseLoadConfigDirSuccess is a simple test that just verifies that
@@ -37,7 +39,7 @@ func TestParserLoadConfigDirSuccess(t *testing.T) {
 			parser := NewParser(nil)
 			path := filepath.Join("testdata/valid-modules", name)
 
-			mod, diags := parser.LoadConfigDir(path)
+			mod, diags := parser.LoadConfigDir(path, RootModuleCallForTesting())
 			if len(diags) != 0 && len(mod.ActiveExperiments) != 0 {
 				// As a special case to reduce churn while we're working
 				// through experimental features, we'll ignore the warning
@@ -103,7 +105,14 @@ func TestParserLoadConfigDirSuccess(t *testing.T) {
 				"mod/" + name: string(src),
 			})
 
-			_, diags := parser.LoadConfigDir("mod")
+			_, diags := parser.LoadConfigDir("mod", NewStaticModuleCall(addrs.RootModule,
+				func(v *Variable) (cty.Value, hcl.Diagnostics) {
+					if !v.Required() {
+						// Allow defaults in this test
+						return v.Default, nil
+					}
+					panic("Variables not configured for this test!")
+				}, "<testing>", ""))
 			if diags.HasErrors() {
 				t.Errorf("unexpected error diagnostics")
 				for _, diag := range diags {
@@ -133,7 +142,7 @@ func TestParserLoadConfigDirWithTests(t *testing.T) {
 			}
 
 			parser := NewParser(nil)
-			mod, diags := parser.LoadConfigDirWithTests(directory, testDirectory)
+			mod, diags := parser.LoadConfigDirWithTests(directory, testDirectory, RootModuleCallForTesting())
 			if len(diags) > 0 { // We don't want any warnings or errors.
 				t.Errorf("unexpected diagnostics")
 				for _, diag := range diags {
@@ -150,7 +159,7 @@ func TestParserLoadConfigDirWithTests(t *testing.T) {
 
 func TestParserLoadConfigDirWithTests_ReturnsWarnings(t *testing.T) {
 	parser := NewParser(nil)
-	mod, diags := parser.LoadConfigDirWithTests("testdata/valid-modules/with-tests", "not_real")
+	mod, diags := parser.LoadConfigDirWithTests("testdata/valid-modules/with-tests", "not_real", RootModuleCallForTesting())
 	if len(diags) != 1 {
 		t.Errorf("expected exactly 1 diagnostic, but found %d", len(diags))
 	} else {
@@ -197,7 +206,7 @@ func TestParserLoadConfigDirFailure(t *testing.T) {
 			parser := NewParser(nil)
 			path := filepath.Join("testdata/invalid-modules", name)
 
-			_, diags := parser.LoadConfigDir(path)
+			_, diags := parser.LoadConfigDir(path, RootModuleCallForTesting())
 			if !diags.HasErrors() {
 				t.Errorf("no errors; want at least one")
 				for _, diag := range diags {
@@ -226,7 +235,7 @@ func TestParserLoadConfigDirFailure(t *testing.T) {
 				"mod/" + name: string(src),
 			})
 
-			_, diags := parser.LoadConfigDir("mod")
+			_, diags := parser.LoadConfigDir("mod", RootModuleCallForTesting())
 			if !diags.HasErrors() {
 				t.Errorf("no errors; want at least one")
 				for _, diag := range diags {